Zips: encode files as unicode
On modern versions of Windows, the filenames encoded with the
default encoding do not correctly extract files with special
characters e.g. french accents. Specify that filenames are to
be Unicode (UTF-8) encoded to try to avoid this. Unfortunately,
Windows zip read/write is flaky so this probably won't handle
every use case, but this should handle more cases than before.